add very basic authentication for card resources

+++ src/main/java/me/tisana/miniblog/web/rest/CardResource.java
@@ -1,27 +1,29 @@
 package me.tisana.miniblog.web.rest;
 
-import me.tisana.miniblog.service.CardService;
-import me.tisana.miniblog.web.rest.errors.BadRequestAlertException;
-import me.tisana.miniblog.service.dto.CardDTO;
-
 import io.github.jhipster.web.util.HeaderUtil;
 import io.github.jhipster.web.util.PaginationUtil;
 import io.github.jhipster.web.util.ResponseUtil;
+import me.tisana.miniblog.domain.Author;
+import me.tisana.miniblog.service.AuthorService;
+import me.tisana.miniblog.service.CardService;
+import me.tisana.miniblog.service.dto.CardDTO;
+import me.tisana.miniblog.web.rest.errors.BadRequestAlertException;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 /**
@@ -34,14 +36,18 @@ public class CardResource {
     private final Logger log = LoggerFactory.getLogger(CardResource.class);
 
     private static final String ENTITY_NAME = "miniBlogCard";
+    private static final String AUTH_REQUIRED_ERROR_KEY = "authentication required";
 
     @Value("${jhipster.clientApp.name}")
     private String applicationName;
 
     private final CardService cardService;
+    private final AuthorService authorService;
 
-    public CardResource(CardService cardService) {
+
+    public CardResource(CardService cardService, AuthorService authorService) {
         this.cardService = cardService;
+        this.authorService = authorService;
     }
 
     /**
@@ -57,6 +63,11 @@ public class CardResource {
         if (cardDTO.getId() != null) {
             throw new BadRequestAlertException("A new card cannot already have an ID", ENTITY_NAME, "idexists");
         }
+
+        if (cardDTO.getAuthorId() == null && cardDTO.getAuthorUsername() == null) {
+            throw new BadRequestAlertException("Author is required", ENTITY_NAME, AUTH_REQUIRED_ERROR_KEY);
+        }
+
         CardDTO result = cardService.save(cardDTO);
         return ResponseEntity.created(new URI("/api/cards/" + result.getId()))
             .headers(HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, result.getId().toString()))
@@ -78,6 +89,21 @@ public class CardResource {
         if (cardDTO.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+
+        if (cardDTO.getAuthorId() == null && cardDTO.getAuthorUsername() == null) {
+            throw new BadRequestAlertException("Author is required", ENTITY_NAME, AUTH_REQUIRED_ERROR_KEY);
+        }
+
+        if (cardDTO.getAuthorPassword() == null) {
+            throw new BadRequestAlertException("Authentication is required", ENTITY_NAME, AUTH_REQUIRED_ERROR_KEY);
+        }
+
+        //check username + password matching
+        Author author = authorService.findOneByUsername(cardDTO.getAuthorUsername());
+        if (author == null || !(author.getPassword().equals(cardDTO.getAuthorPassword()))) {
+            throw new BadRequestAlertException("Invalid Credential", ENTITY_NAME, AUTH_REQUIRED_ERROR_KEY);
+        }
+
         CardDTO result = cardService.save(cardDTO);
         return ResponseEntity.ok()
             .headers(HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, cardDTO.getId().toString()))
@@ -118,8 +144,22 @@ public class CardResource {
      * @return the {@link ResponseEntity} with status {@code 204 (NO_CONTENT)}.
      */
     @DeleteMapping("/cards/{id}")
-    public ResponseEntity<Void> deleteCard(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCard(@PathVariable Long id, @RequestParam Map<String, String> params) {
         log.debug("REST request to delete Card : {}", id);
+
+        String username = params.get("authorUsername");
+        String password = params.get("password");
+
+        if (StringUtils.isBlank(username) || StringUtils.isBlank(password)) {
+            throw new BadRequestAlertException("Authentication is required", ENTITY_NAME, AUTH_REQUIRED_ERROR_KEY);
+        }
+
+        //check username + password matching
+        Author author = authorService.findOneByUsername(username);
+        if (author == null || !(author.getPassword().equals(password))) {
+            throw new BadRequestAlertException("Invalid Credential", ENTITY_NAME, AUTH_REQUIRED_ERROR_KEY);
+        }
+
         cardService.delete(id);
         return ResponseEntity.noContent().headers(HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString())).build();
     }
